Move datadog-core contents to a package named core
