+++ app/app/src/main/kotlin/com/hedvig/app/startup/DatadogInitializerImpl.kt
@@ -1,7 +1,7 @@
 package com.hedvig.app.startup
 
 import androidx.startup.Initializer
-import com.hedvig.android.datadog.DatadogInitializer
+import com.hedvig.android.datadog.core.DatadogInitializer
 import com.hedvig.app.feature.di.KoinInitializer
 
 @Suppress("unused") // Used in /app/src/main/AndroidManifest.xml
