+++ app/datadog/datadog-core/src/main/kotlin/com/hedvig/android/datadog/core/OkHttpDatadogUtil.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.datadog
+package com.hedvig.android.datadog.core
 
 import com.datadog.android.DatadogEventListener
 import com.datadog.android.DatadogInterceptor
