+++ app/datadog/datadog-core/src/main/kotlin/com/hedvig/android/datadog/core/di/DatadogModule.kt
@@ -1,7 +1,7 @@
-package com.hedvig.android.datadog.di
+package com.hedvig.android.datadog.core.di
 
 import com.hedvig.android.auth.event.AuthEventListener
-import com.hedvig.android.datadog.memberid.DatadogMemberIdUpdatingAuthEventListener
+import com.hedvig.android.datadog.core.memberid.DatadogMemberIdUpdatingAuthEventListener
 import io.opentracing.Tracer
 import io.opentracing.util.GlobalTracer
 import org.koin.dsl.bind
